% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm.R
\name{dtm}
\alias{dtm}
\title{Wrapper for dtm}
\usage{
dtm(args = "--help")
}
\arguments{
\item{args}{parameters passed to the main dtm program}
}
\description{
Wraps the dtm executable. WIP.
}
\examples{
# Run help page
dtm("--help")
}
